% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlestick_plotly.R
\name{candlestick_plotly}
\alias{candlestick_plotly}
\title{Create a candlestick chart with Plotly}
\usage{
candlestick_plotly(ochl_df, id)
}
\arguments{
\item{ochl_df}{A data.frame containing open, close, high and low columns}

\item{id}{'The name of the ticker included in the column names}
}
\value{
A Plotly object
}
\description{
Create a candletick plot using a df with the same structure as the xts object returned by \code{quantmod::getSymbols()}
}
