% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adjusted_prices.R
\name{get_adjusted_prices}
\alias{get_adjusted_prices}
\title{Get adjusted stock prices.}
\usage{
get_adjusted_prices(tickers)
}
\arguments{
\item{tickers}{a vector of strings containing valid ticker names}
}
\value{
an xts object
}
\description{
Download the adjusted price of multiple tickers from yahoo and merge them
into one xts
}
\examples{
get_adjusted_prices(c('ATVI', 'TTWO'))
}
