% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_finviz.R
\name{get_table_finviz}
\alias{get_table_finviz}
\title{Get table from finviz}
\usage{
get_table_finviz(ticker)
}
\arguments{
\item{ticker}{a valid ticker name}
}
\value{
tibble
}
\description{
Download the table with indicators located below the graph in finviz
}
\examples{
get_table_finviz('ATVI')
}
