% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_thead_kable.R
\name{no_thead_kable}
\alias{no_thead_kable}
\title{A kable with no thead}
\usage{
no_thead_kable(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
html table
}
\description{
A kable with no thead
}
\examples{
no_thead_kable(mtcars)
}
