% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickStock.R
\name{pickStockServer}
\alias{pickStockServer}
\title{Server function for pickStock module}
\usage{
pickStockServer(id, choices)
}
\arguments{
\item{id}{The server id}

\item{choices}{A character vector for the choices}
}
\value{
a reactive character vector
}
\description{
Server function for pickStock module
}
\examples{
pickStockServer('my_id', c('choice1', 'choice2'))
}
