% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickStock.R
\name{pickStockUI}
\alias{pickStockUI}
\title{UI for pickStock module}
\usage{
pickStockUI(id)
}
\arguments{
\item{id}{a name to supply to \code{NS()}}
}
\value{
taglist
}
\description{
Make a manual or random pick from a selectizeInput containing some choices
}
