% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dev_ret.R
\name{plot_dev_ret}
\alias{plot_dev_ret}
\title{Scatterplot with deviations from the mean}
\usage{
plot_dev_ret(xts_ret, stdev)
}
\arguments{
\item{xts_ret}{xts object from \code{quantmod::getSymbols()}}

\item{stdev}{a standart deviation}
}
\value{
ggplot
}
\description{
Make a plot with the returns marked as green is they are above one standart
deviation and red is vice-versa.
}
