% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{returnsServer}
\alias{returnsServer}
\title{Server function for returns module}
\usage{
returnsServer(id, returns_data)
}
\arguments{
\item{id}{}

\item{returns_data}{}
}
\description{
Server function for returns module
}
