% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmod.R
\name{riskmodServer}
\alias{riskmodServer}
\title{Server function for riskmod module.}
\usage{
riskmodServer(id, returns_data)
}
\arguments{
\item{id}{a name for the id}

\item{returns_data}{an xts object containing one column of returns}
}
\description{
Server function for riskmod module.
}
