% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run stock application.}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional arguments to pass to \code{shinyApp()}}
}
\description{
This application let's the user download information about sp500 tickers and
take a look at different metrics and indicators.
}
