% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickerInfo.R
\name{tickerInfoServer}
\alias{tickerInfoServer}
\title{Server for tickerInfoServer}
\usage{
tickerInfoServer(id)
}
\arguments{
\item{id}{a name for the id}
}
\value{

}
\description{
Server for tickerInfoServer
}
