% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_to_tibble.R
\name{xts_to_tibble}
\alias{xts_to_tibble}
\title{Convert an xts to a tibble}
\usage{
xts_to_tibble(xts_obj)
}
\arguments{
\item{xts_obj}{xts object}
}
\value{
tibble
}
\description{
Convert an xts to a tibble
}
